#change from double to int

+++ MusicAndReaction.Server/Database/ReactionDTO.cs
@@ -2,5 +2,5 @@
 public class ReactionDTO
 {
     public int TrackId { get; set; }
-    public double ReactionTime { get; set; }
+    public int ReactionTime { get; set; }
 }
