+++ MusicAndReaction.Server/Services/DatabaseService.cs
@@ -5,8 +5,8 @@ public class DatabaseService(AppDbContext context) : IDatabaseService
 {
     #region GetData
     public async Task<List<ReactionDTO>> GetAllDataList() => await context.UserReactions.ToDto().ToListAsync();
-    public async Task<List<double>> GetDataForMusic(int trackid) => await context.UserReactions.Where(x => x.TrackId == trackid).Select(x=>x.ReactionTime).ToListAsync();
-    public async Task<List<double>> GetOrderByForMusic(int trackid) => await context.UserReactions.Where(x => x.TrackId == trackid).OrderBy(x => x.ReactionTime).Select(x => x.ReactionTime).ToListAsync();
+    public async Task<List<int>> GetDataForMusic(int trackid) => await context.UserReactions.Where(x => x.TrackId == trackid).Select(x=>x.ReactionTime).ToListAsync();
+    public async Task<List<int>> GetOrderByForMusic(int trackid) => await context.UserReactions.Where(x => x.TrackId == trackid).OrderBy(x => x.ReactionTime).Select(x => x.ReactionTime).ToListAsync();
     public async Task<List<ReactionDTO>> GetOrderByDataList() => await context.UserReactions.OrderBy(x => x.ReactionTime).ToDto().ToListAsync();
     #endregion
