+++ MusicAndReaction.Server/Services/IDatabaseService.cs
@@ -4,7 +4,7 @@ public interface IDatabaseService
 {
     Task AddReaction(ReactionDTO newReaction);
     Task<List<ReactionDTO>> GetAllDataList();
-    Task<List<double>> GetDataForMusic(int trackid);
-    Task<List<double>> GetOrderByForMusic(int trackid);
+    Task<List<int>> GetDataForMusic(int trackid);
+    Task<List<int>> GetOrderByForMusic(int trackid);
     Task<List<ReactionDTO>> GetOrderByDataList();
 }
